Extract ws close status retrieval to common
